third_party/blink/web_tests/external/wpt/css/css-typed-om/stylevalue-subclasses/cssColorValue.html

<!doctype html>
<meta charset="utf-8">
<title>CSSColorValue Tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#csscolorvalue">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<script>
'use strict';

const TEST_COLORS = [
  {name: "magenta", rgb: [1, 0, 1], hsl: [300, 1, 0.5], hwb: [300, 0, 0]},
  {name: "dark cyan", rgb: [0, 0.545, 0.545], hsl: [180, 1, 0.2725], hwb: [180, 0, 0.455]},
  {name: "light goldenrod yellow", rgb: [1, 1, 0.82], hsl: [60, 1, 0.91], hwb: [60, 0.82, 0]},
  {name: "medium purple", rgb: [0.58, 0.44, 0.86], hsl: [260, 0.5977, 0.649], hwb: [260, 0.44, 0.14]},
]

for (const color of TEST_COLORS) {
  const hsl_color_degrees = new CSSHSL(CSS.deg(color.hsl[0]), color.hsl[1], color.hsl[2])
  const hsl_color_radians = new CSSHSL(CSS.rad(color.hsl[0]/360 * 2 * Math.PI), color.hsl[1], color.hsl[2])
  const hsl_color_gradians = new CSSHSL(CSS.grad(color.hsl[0]/360 * 400), color.hsl[1], color.hsl[2])
  const hsl_attributes = ['h', 's', 'l', 'alpha']

  const hwb_color_degrees = new CSSHWB(CSS.deg(color.hwb[0]), color.hwb[1], color.hwb[2])
  const hwb_color_radians = new CSSHWB(CSS.rad(color.hwb[0]/360 * 2 * Math.PI), color.hwb[1], color.hwb[2])
  const hwb_color_gradians = new CSSHWB(CSS.grad(color.hwb[0]/360 * 400), color.hwb[1], color.hwb[2])
  const hwb_attributes = ['h', 'w', 'b', 'alpha']

  const rgb_color = new CSSRGB(color.rgb[0], color.rgb[1], color.rgb[2])
  const rgb_attributes = ['r', 'g', 'b', 'alpha']

  // Test conversion from CSSHSL to others.
  test(() => {
    const hsl_to_rgb = hsl_color_degrees.toRGB();
    for (const attr of rgb_attributes) {
      assert_color_channel_approx_equals(hsl_to_rgb[attr], rgb_color[attr]);
    }
  }, `Converting HSL using degrees to RGB works for ${color.name}.`);

  test(() => {
    const hsl_to_rgb = hsl_color_radians.toRGB();
    for (const attr of rgb_attributes) {
      assert_color_channel_approx_equals(hsl_to_rgb[attr], rgb_color[attr]);
    }
  }, `Converting HSL using radians to RGB works for ${color.name}.`);

  test(() => {
    const hsl_to_rgb = hsl_color_gradians.toRGB();
    for (const attr of rgb_attributes) {
      assert_color_channel_approx_equals(hsl_to_rgb[attr], rgb_color[attr]);
    }
  }, `Converting HSL using gradians to RGB works for ${color.name}.`);

  test(() => {
    const hsl_to_hwb = hsl_color_degrees.toHWB();
    for (const attr of hwb_attributes) {
      assert_color_channel_approx_equals(hwb_color_degrees[attr], hsl_to_hwb[attr]);
    }
  }, `Converting HSL to HWB works for ${color.name}.`);

  // Test conversion from CSSHWB to others.
  test(() => {
    const hwb_to_rgb = hwb_color_degrees.toRGB();
    for (const attr of rgb_attributes) {
      assert_color_channel_approx_equals(hwb_to_rgb[attr], rgb_color[attr]);
    }
  }, `Converting HWB using degrees to RGB works for ${color.name}.`);

  test(() => {
    const hwb_to_rgb = hwb_color_radians.toRGB();
    for (const attr of rgb_attributes) {
      assert_color_channel_approx_equals(hwb_to_rgb[attr], rgb_color[attr]);
    }
  }, `Converting HWB using radians to RGB works for ${color.name}.`);

  test(() => {
    const hwb_to_rgb = hwb_color_gradians.toRGB();
    for (const attr of rgb_attributes) {
      assert_color_channel_approx_equals(hwb_to_rgb[attr], rgb_color[attr]);
    }
  }, `Converting HWB using gradians to RGB works for ${color.name}.`);

  test(() => {
    const hwb_to_hsl = hwb_color_degrees.toHSL();
    for (const attr of hsl_attributes) {
      assert_color_channel_approx_equals(hsl_color_degrees[attr], hwb_to_hsl[attr]);
    }
  }, `Converting HWB to HSL works for ${color.name}.`);

  // Test conversion from CSSRGB to others
  test(() => {
    const rgb_to_hsl = rgb_color.toHSL();
    for (const attr of hsl_attributes) {
      assert_color_channel_approx_equals(hsl_color_degrees[attr], rgb_to_hsl[attr]);
    }
  }, `Converting RGB to HSL works for ${color.name}.`);

  test(() => {

    const rgb_to_hwb = rgb_color.toHWB();
    for (const attr of hwb_attributes) {
      assert_color_channel_approx_equals(hwb_color_degrees[attr], rgb_to_hwb[attr]);
    }
  }, `Converting RGB to HWB works for ${color.name}.`);
}
</script>
